Add Midje test framework dependency
Instead of putting the (unit) test files in a separate test directory I choose
to put the test file next to the file (namespace) it is testing. By doing this
we do not have to duplicate the src directory structure into the test directory
anymore. Furthermore, I think it is handy to have the tests near by. And if you
see these tests as kind of a specification (by example) of your code, why not
put them together.

The :jar-exclusions and :uberjar-exclusions were added to the project.clj to
exclude the test files from the (uber)jars. I seems that these exclusions
weren't effective for the :uberjar :aot resulting in a not found of the midje
(dev) dependency. To solve this, the namespace to compile ahead of time
(nand-board.core) is now explicitly specified. The assumption is that not only
this namespace is compiled but also all the namespaces it is depending on.