Remove :val's from board
The values of the pins will be stored in some state object, separate from the
board. It will then probably be easier for the (yet to be implemented)
simulator to maintain this state and to do reevaluations (when a new wire is
connected for example) and such. Furthermore, I think it's a good idea to let
the board only contains the 'structure' of the board (the gates with their
pins and how they are wired).