Introduce :id naming convention for domain data entities
The board data structure uses id's to cross reference domain data entities
(gates, pins, wires). In my opinion this should be treated as a technical
detail of the implementation. The rest of the application like the Simulator
should not be aware of this detail, because it is then difficult to make
changes or to refactor the data structure.

I am planning to refactor the application to work with the domain data
entities and to not use for example the value of the :input-pin-ids of a gate
directly. The board will then instead define functions to get the input pins
for a given gate.

Having said that, the application may still want to identify a domain data
entity by an identifier. For example the Simulator wants to have a map of
values for the pins on the board. Using the pin entity itself as the key of a
value in this map does not feel okay. The pin is immutable. In the current
implementation adding a wire to the board for this pin will result in a new
pin (and a new board). But from the viewpoint of the Simulator I want to treat
it as the same pin. You may argue that it is/has an identity then and use Refs.
But leaving the path of immutability does not feel okay either. Therefore it
seems to me that it is a good idea to introduce a name convention for the id
of a domain data entity that may be used throughout the application. In other
words, feel free to use the :id, but do not use the :input-pin-ids,
:output-pin-id, :wire-ids, etc.